requirements.txt
Flask==3.0.3
redis==4.6.0
pymongo==4.6.3
pydantic==2.4.0
requests~=2.32.0
Pillow~=10.3.0
markupsafe~=2.1.3
gevent~=24.2.1
flask_caching
flask-cors~=5.0.0
xmltodict~=0.13.0
flask-caching~=2.3.0
blinker~=1.8.2
cos-python-sdk-v5
